G-API GExecutor internal interface
